// src/types/errors.ts
export interface Errors {
    email?: string;
    password?: string;
    passwordConfirm?: string;
    name?: string;
    surname?: string;
    cpf?: string;
    cellphone?: string;
    apelido?: string;
    cep?: string;
    rua?: string;
    cidade?: string;
    bairro?: string;
    estado?: string;
    num?: string;
  }